backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.14
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiofiles==25.1.0                    # via -r requirements.in
annotated-doc==0.0.4                # via fastapi
annotated-types==0.7.0              # via pydantic
anyio==4.12.1                       # via httpx, starlette
asyncpg==0.31.0                     # via -r requirements.in
certifi==2026.1.4                   # via httpcore, httpx
click==8.3.1                        # via uvicorn
coverage[toml]==7.13.3              # via pytest-cov
fastapi==0.135.1                    # via -r requirements.in
greenlet==3.3.2                     # via sqlalchemy
h11==0.16.0                         #   via httpcore, uvicorn
httpcore==1.0.9                     # via httpx
httpx==0.28.1                       # via -r requirements.in
idna==3.11                          # via anyio httpx
iniconfig==2.3.0                    # via pytest
jinja2==3.1.6                       # via pytest-html
markupsafe==3.0.3                   # via jinja2
packaging==26.0                     # via pytest
pluggy==1.6.0                       # via pytest  pytest-cov
py-cpuinfo==9.0.0                   # via pytest-benchmark
pydantic==2.12.5                    # via -r requirements.in fastapi pydantic-settings
pydantic-core==2.41.5               # via -r requirements.in pydantic
pydantic-settings==2.13.1           # via -r requirements.in
pygments==2.19.2                    # via pytest
pytest==9.0.2                       # via -r requirements.in pytest-asyncio, pytest-benchmark,pytest-cov, pytest-html, pytest-metadata
pytest-asyncio==1.3.0               # via -r requirements.in
pytest-benchmark==5.2.3             # via -r requirements.in
pytest-cov==7.0.0                   # via -r requirements.in
pytest-html==4.2.0                  # via -r requirements.in
pytest-metadata==3.1.1              # via -r requirements.in pytest-html
python-dotenv==1.2.1                # via -r requirements.in pydantic-settings
simplified-logging==1.0.13.7        # via -r requirements.in
sqlalchemy==2.0.48                  # via -r requirements.in
starlette==0.52.1                   # via fastapi
typing-extensions==4.15.0           # via fastapi pydantic pydantic-core sqlalchemy typing-inspection
typing-inspection==0.4.2            # via fastapi pydantic pydantic-settings
uvicorn==0.41.0                     # via -r requirements.in
